api/requirements.txt
-e git+https://github.com/GaNiziolek/base-fullstack.git@72417a5b7404426b26b9fd2714e61c51c7b15d33#egg=api&subdirectory=api
arrow==1.2.3
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colander==1.8.3
ColanderAlchemy==0.3.4
cookiecutter==2.1.1
cornice==6.0.1
greenlet==2.0.1
hupper==1.10.3
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
PasteDeploy==3.0.1
plaster==1.0
plaster-pastedeploy==0.7
psycopg2-binary==2.9.5
pyramid==2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.43
text-unidecode==1.3
translationstring==1.4
urllib3==1.26.12
venusian==3.0.0
waitress==2.1.2
WebOb==1.8.7
zope.deprecation==4.4.0
zope.interface==5.5.0
